dom/inputmethod/mochitest/test_bug978918.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=978918
-->
<head>
  <title>Basic test for InputMethod API.</title>
  <script type="application/javascript;version=1.7" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript;version=1.7" src="inputmethod_common.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=978918">Mozilla Bug 978918</a>
<p id="display"></p>
<pre id="test">
<script class="testbody" type="application/javascript;version=1.7">

// The input context.
var gContext = null;

inputmethod_setup(function() {
  runTest();
});

function runTest() {
  let im = navigator.mozInputMethod;

  im.oninputcontextchange = function() {
    ok(true, 'inputcontextchange event was fired.');
    im.oninputcontextchange = null;

    gContext = im.inputcontext;
    if (!gContext) {
      ok(false, 'Should have a non-null inputcontext.');
      inputmethod_cleanup();
      return;
    }

    test_setSelectionRange();
  };

  // Set current page as an input method.
  SpecialPowers.wrap(im).setActive(true);

  let iframe = document.createElement('iframe');
  iframe.src = 'file_test_sms_app.html';
  iframe.setAttribute('mozbrowser', true);
  document.body.appendChild(iframe);
}

function test_setSelectionRange() {
  gContext.setSelectionRange(0, 100).then(function() {
    is(gContext.selectionStart, 0, 'selectionStart was set successfully.');
    is(gContext.selectionEnd, 5, 'selectionEnd was set successfully.');
    test_replaceSurroundingText();
  }, function(e) {
    ok(false, 'setSelectionRange failed:' + e.name);
    inputmethod_cleanup();
  });
}

function test_replaceSurroundingText() {
  // Replace 'Httvb' with 'Hito'.
  gContext.replaceSurroundingText('Hito', 0, 100).then(function() {
    ok(true, 'replaceSurroundingText finished');
    inputmethod_cleanup();
  }, function(e) {
    ok(false, 'replaceSurroundingText failed: ' + e.name);
    inputmethod_cleanup();
  });
}

</script>
</pre>
</body>
</html>
